# rest/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: enviroCar REST API
  description: enviroCar REST API
  version: "1.0.0"
  contact:
    email: info@envirocar.org
    name: "52\xB0North GmbH"
    url: https://52north.org
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://envirocar.org/conditions.html?lng=en
servers: [ ]
...
paths:
...
  '/api-docs':
    get:
      operationId: apiDocs
      summary: Get the OpenAPI specification of this service.
      description: Get the OpenAPI specification of this service.
      tags:
        - General
      responses:
        '200':
          $ref: '#/components/responses/ApiDocs'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
  '/phenomenons':
    get:
      operationId: getPhenomenons
      summary: Get the phenomenon listing.
      description: Get the phenomenon listing.
      tags:
        - Phenomenons
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/range'
      responses:
        '200':
          $ref: '#/components/responses/Phenomenons'
        '206':
          $ref: '#/components/responses/Phenomenons'
        '400':
          $ref: '#/components/responses/BadRequest'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createPhenomenon
      summary: Create new phenomenon.
      description: Create new phenomenon.
      tags:
        - Phenomenons
      security:
        - basicAuth: [ ]
      requestBody:
        $ref: '#/components/requestBodies/CreatePhenomenon'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '451':
          $ref: '#/components/responses/UnavailableForLegalReasons'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
...
  '/sensors':
    get:
      operationId: getSensors
      summary: Get the Sensor listing.
      description: Get the Sensor listing.
      tags:
        - Sensors
      parameters:
        - $ref: '#/components/parameters/sensorType'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/range'
      responses:
        '200':
          $ref: '#/components/responses/Sensors'
        '206':
          $ref: '#/components/responses/Sensors'
        '400':
          $ref: '#/components/responses/BadRequest'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createSensor
      summary: Create a new Sensor.
      description: Create a new Sensor.
      tags:
        - Sensors
      requestBody:
        $ref: '#/components/requestBodies/CreateSensor'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '451':
          $ref: '#/components/responses/UnavailableForLegalReasons'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
...
  '/tracks':
    get:
      operationId: getTracks
      summary: Get the list of tracks.
      description: Get the list of tracks.
      tags:
        - Tracks
      parameters:
        - $ref: '#/components/parameters/bbox'
        - $ref: '#/components/parameters/nearPoint'
        - $ref: '#/components/parameters/after'
        - $ref: '#/components/parameters/before'
        - $ref: '#/components/parameters/begins'
        - $ref: '#/components/parameters/begunBy'
        - $ref: '#/components/parameters/contains'
        - $ref: '#/components/parameters/during'
        - $ref: '#/components/parameters/endedBy'
        - $ref: '#/components/parameters/ends'
        - $ref: '#/components/parameters/equals'
        - $ref: '#/components/parameters/meets'
        - $ref: '#/components/parameters/metBy'
        - $ref: '#/components/parameters/overlapped'
        - $ref: '#/components/parameters/overlaps'
        - $ref: '#/components/parameters/status'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/range'
      responses:
        '200':
          $ref: '#/components/responses/Tracks'
        '206':
          $ref: '#/components/responses/Tracks'
        '400':
          $ref: '#/components/responses/BadRequest'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createTrack
      summary: Create a new track.
      description: Create a new track.
      tags:
        - Tracks
      security:
        - basicAuth: [ ]
      requestBody:
        $ref: '#/components/requestBodies/CreateTrack'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '451':
          $ref: '#/components/responses/UnavailableForLegalReasons'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
...
  '/measurements':
    get:
      operationId: getMeasurements
      summary: Get the list of measurements.
      description: Get the list of measurements.
      tags:
        - Measurements
      parameters:
        - $ref: '#/components/parameters/bbox'
        - $ref: '#/components/parameters/nearPoint'
        - $ref: '#/components/parameters/after'
        - $ref: '#/components/parameters/before'
        - $ref: '#/components/parameters/begins'
        - $ref: '#/components/parameters/begunBy'
        - $ref: '#/components/parameters/contains'
        - $ref: '#/components/parameters/during'
        - $ref: '#/components/parameters/endedBy'
        - $ref: '#/components/parameters/ends'
        - $ref: '#/components/parameters/equals'
        - $ref: '#/components/parameters/meets'
        - $ref: '#/components/parameters/metBy'
        - $ref: '#/components/parameters/overlapped'
        - $ref: '#/components/parameters/overlaps'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/range'
      responses:
        '200':
          $ref: '#/components/responses/Measurements'
        '206':
          $ref: '#/components/responses/Measurements'
        '400':
          $ref: '#/components/responses/BadRequest'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createMeasurement
      summary: Create a new measurement.
      description: Create a new measurement.
      tags:
        - Measurements
      security:
        - basicAuth: [ ]
      requestBody:
        $ref: '#/components/requestBodies/CreateMeasurement'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '451':
          $ref: '#/components/responses/UnavailableForLegalReasons'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
...
  '/privacyStatements':
    get:
      operationId: getPrivacyStatements
      summary: Get the list of privacy statements.
      description: Get the list of privacy statements.
      tags:
        - Terms
      responses:
        '200':
          $ref: '#/components/responses/PrivacyStatements'
        '400':
          $ref: '#/components/responses/BadRequest'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
...
  '/users':
    get:
      operationId: getUsers
      summary: Get the users.
      description: Get the users.
      tags:
        - Users
      security:
        - basicAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/range'
      responses:
        '200':
          $ref: '#/components/responses/Users'
        '206':
          $ref: '#/components/responses/Users'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createUser
      summary: Create a new user.
      description: Create a new user.
      tags:
        - Users
      security:
        - basicAuth: [ ]
      requestBody:
        $ref: '#/components/requestBodies/CreateSensor'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '451':
          $ref: '#/components/responses/UnavailableForLegalReasons'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
...
  '/users/{userId}/userStatistic':
    parameters:
      - $ref: '#/components/parameters/userId'
    get:
      operationId: getUserStatistics
      summary: Get the user statistics.
      description: Get the user statistics.
      tags:
        - Users
        - Statistics
      security:
        - basicAuth: [ ]
      responses:
        '200':
          $ref: '#/components/responses/UserStatistic'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
...
  '/users/{userId}/fuelings':
    parameters:
      - $ref: '#/components/parameters/userId'
    get:
      operationId: getFuelings
      summary: Get the fuelings for the user.
      description: Get the fuelings for the user.
      tags:
        - Fuelings
        - Users
      security:
        - basicAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/after'
        - $ref: '#/components/parameters/before'
        - $ref: '#/components/parameters/begins'
        - $ref: '#/components/parameters/begunBy'
        - $ref: '#/components/parameters/contains'
        - $ref: '#/components/parameters/during'
        - $ref: '#/components/parameters/endedBy'
        - $ref: '#/components/parameters/ends'
        - $ref: '#/components/parameters/equals'
        - $ref: '#/components/parameters/meets'
        - $ref: '#/components/parameters/metBy'
        - $ref: '#/components/parameters/overlapped'
        - $ref: '#/components/parameters/overlaps'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/range'
      responses:
        '200':
          $ref: '#/components/responses/Fuelings'
        '206':
          $ref: '#/components/responses/Fuelings'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createFueling
      summary: Create a new fueling for the user.
      description: Create a new fueling for the user.
      tags:
        - Fuelings
        - Users
      security:
        - basicAuth: [ ]
      requestBody:
        $ref: '#/components/requestBodies/CreateFueling'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
...
  '/groups':
    get:
      operationId: getGroups
      summary: Get the groups.
      description: Get the groups.
      parameters:
        - $ref: '#/components/parameters/q'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/range'
      tags:
        - Groups
      security:
        - basicAuth: [ ]
      responses:
        '200':
          $ref: '#/components/responses/Groups'
        '206':
          $ref: '#/components/responses/Groups'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createGroup
      summary: Create the group.
      description: Create the group.
      tags:
        - Users
        - Groups
      security:
        - basicAuth: [ ]
      requestBody:
        $ref: '#/components/requestBodies/CreateGroup'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
...
  '/badges':
    get:
      operationId: getBadges
      summary: Get the list of badges.
      description: Get the list of badges.
      tags:
        - Badges
      responses:
        '200':
          $ref: '#/components/responses/Badges'
        '400':
          $ref: '#/components/responses/BadRequest'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
  '/announcements':
    get:
      operationId: getAnnouncements
      summary: Get the list of announcements.
      description: Get the list of announcements.
      tags:
        - Announcements
      responses:
        '200':
          $ref: '#/components/responses/Announcements'
        '400':
          $ref: '#/components/responses/BadRequest'
        '5XX':
          $ref: '#/components/responses/InternalServerError'
components:
  parameters:
    after:
      description: The time instant or interval to be applied with the `after` operator.
      in: query
      name: after
      required: false
      schema:
        $ref: '#/components/schemas/TemporalFilter'
    bbox:
      description: The bounding box the requested tracks have to intersect with.
      explode: false
      in: query
      name: bbox
      required: false
      style: form
      schema:
        items:
          type: number
        maxItems: 4
        minItems: 4
        type: array
    before:
      description: The time instant or interval to be applied with the `before` operator.
      in: query
      name: before
      required: false
      schema:
        $ref: '#/components/schemas/TemporalFilter'
    begins:
      description: The time instant or interval to be applied with the `begins` operator.
      in: query
      name: begins
      required: false
      schema:
        $ref: '#/components/schemas/TemporalFilter'
    begunBy:
      description: The time instant or interval to be applied with the `begunBy` operator.
      in: query
      name: begunBy
      required: false
      schema:
        $ref: '#/components/schemas/TemporalFilter'
    contains:
      description: The time instant or interval to be applied with the `contains`
        operator.
      in: query
      name: contains
      required: false
      schema:
        $ref: '#/components/schemas/TemporalFilter'
    during:
      description: The time instant or interval to be applied with the `during` operator.
      in: query
      name: during
      required: false
      schema:
        $ref: '#/components/schemas/TemporalFilter'
    endedBy:
      description: The time instant or interval to be applied with the `endedBy` operator.
      in: query
      name: endedBy
      required: false
      schema:
        $ref: '#/components/schemas/TemporalFilter'
    ends:
      description: The time instant or interval to be applied with the `ends` operator.
      in: query
      name: ends
      required: false
      schema:
        $ref: '#/components/schemas/TemporalFilter'
    equals:
      description: The time instant or interval to be applied with the `equals` operator.
      in: query
      name: equals
      required: false
      schema:
        $ref: '#/components/schemas/TemporalFilter'
...
    meets:
      description: The time instant or interval to be applied with the `meets` operator.
      in: query
      name: meets
      required: false
      schema:
        $ref: '#/components/schemas/TemporalFilter'
    metBy:
      description: The time instant or interval to be applied with the `metBy` operator.
      in: query
      name: metBy
      required: false
      schema:
        $ref: '#/components/schemas/TemporalFilter'
    nearPoint:
      description: The point to which the distance can not fall below a certain threshold.
      explode: false
      in: query
      name: nearPoint
      required: false
      schema:
        items:
          type: number
        maxItems: 3
        minItems: 3
        type: array
      style: form
    overlapped:
      description: The time instant or interval to be applied with the `overlapped`
        operator.
      in: query
      name: overlapped
      required: false
      schema:
        $ref: '#/components/schemas/TemporalFilter'
    overlaps:
      description: The time instant or interval to be applied with the `overlaps`
        operator.
      in: query
      name: overlaps
      required: false
      schema:
        $ref: '#/components/schemas/TemporalFilter'
    status:
      in: query
      name: status
      required: false
      schema:
        $ref: 'schema/track-status.json'
    limit:
      description: The number of entities per page.
      in: query
      name: limit
      required: false
      schema:
        type: integer
        default: 100
    page:
      description: The page of entities to request.
      in: query
      name: page
      required: false
      schema:
        type: integer
        default: 1
    q:
      name: q
      description: The term to search for.
      in: query
      schema:
        type: string
    range:
      description: The range of entities to request.
      in: header
      name: Range
      required: false
      style: simple
      explode: true
      schema:
        type: object
        properties:
          items:
            type: string
            pattern: '^\d+-\d+$'
...
    userId:
      description: The user id.
      in: path
      name: userId
      required: true
      schema:
        type: string
...
    sensorType:
      description: The Sensor type.
      in: query
      name: type
      required: false
      schema:
        type: string
...
  headers:
    ContentRange:
      style: simple
      explode: true
      schema:
        type: string
        pattern: '^items\s+\d+-\d+/(\d+|\*)$'
    Link:
      style: simple
      explode: true
      schema:
        type: object
        properties:
          items:
            type: string
            pattern: '^<.*>\s?;\s?rel=(first|last|next|prev|service-desc)\s?;\s?type=.+/.+'
    Location:
      schema:
        type: string
        format: uri
    WWW-Authenticate:
      schema:
        type: string
  requestBodies:
    CreateMeasurement:
      description: The measurement to create.
      required: true
      content:
        application/json:
          schema:
            $ref: 'schema/measurement-create.json'
    CreatePhenomenon:
      description: The phenomenon to create.
      required: true
      content:
        application/json:
          schema:
            $ref: 'schema/phenomenon-create.json'
    CreateSensor:
      description: The sensor to create.
      required: true
      content:
        application/json:
          schema:
            $ref: 'schema/sensor-create.json'
    CreateTrack:
      description: The track to create.
      required: true
      content:
        application/json:
          schema:
            $ref: 'schema/track-create.json'
...
    CreateFueling:
      description: The fueling to create.
      required: true
      content:
        application/json:
          schema:
            $ref: 'schema/fueling-create.json'
...
  responses:
...
    Created:
      description: The entity was created.
      headers:
        Location:
          $ref: '#/components/headers/Location'
        Link:
          $ref: '#/components/headers/Link'
...
    BadRequest:
      description: The request was malformed.
      content:
        application/json:
          schema:
            $ref: 'schema/exception.json'
      headers:
        Link:
          $ref: '#/components/headers/Link'
    Forbidden:
      description: The operation is not allowed by current user.
      content:
        application/json:
          schema:
            $ref: 'schema/exception.json'
      headers:
        Link:
          $ref: '#/components/headers/Link'
    InternalServerError:
      description: An internal server error occured.
      content:
        application/json:
          schema:
            $ref: 'schema/exception.json'
      headers:
        Link:
          $ref: '#/components/headers/Link'
    NotFound:
      description: The Resource could not be found.
      content:
        application/json:
          schema:
            $ref: 'schema/exception.json'
      headers:
        Link:
          $ref: '#/components/headers/Link'
    UnavailableForLegalReasons:
      description: The operation requieres acceptance of the terms of use and privacy statement.
      content:
        application/json:
          schema:
            $ref: 'schema/exception.json'
      headers:
        Link:
          $ref: '#/components/headers/Link'
    UnauthorizedError:
      description: Authentication information is missing or invalid
      content:
        application/json:
          schema:
            $ref: 'schema/exception.json'
      headers:
        Link:
          $ref: '#/components/headers/Link'
        WWW-Authenticate:
          $ref: '#/components/headers/WWW-Authenticate'
    Phenomenons:
      description: The phenomenons.
      content:
        application/json:
          schema:
            $ref: 'schema/phenomenons.json'
        text/yaml:
          schema:
            $ref: 'schema/phenomenons.json'
      headers:
        Link:
          $ref: '#/components/headers/Link'
        Content-Range:
          $ref: '#/components/headers/ContentRange'
...
    Sensors:
      description: The Sensors.
      content:
        application/json:
          schema:
            $ref: 'schema/sensors.json'
        text/yaml:
          schema:
            $ref: 'schema/sensors.json'
      headers:
        Link:
          $ref: '#/components/headers/Link'
        Content-Range:
          $ref: '#/components/headers/ContentRange'
...
    UserStatistic:
      description: The user statistic.
      content:
        application/json:
          schema:
            $ref: 'schema/userStatistic.json'
        text/yaml:
          schema:
            $ref: 'schema/userStatistic.json'
      headers:
        Link:
          $ref: '#/components/headers/Link'
...
    Tracks:
      description: The tracks.
      content:
        application/json:
          schema:
            $ref: 'schema/tracks.json'
        text/yaml:
          schema:
            $ref: 'schema/tracks.json'
      headers:
        Link:
          $ref: '#/components/headers/Link'
        Content-Range:
          $ref: '#/components/headers/ContentRange'
...
    Measurements:
      description: The measurements.
      content:
        application/json:
          schema:
            $ref: 'schema/measurements.json'
        text/yaml:
          schema:
            $ref: 'schema/measurements.json'
      headers:
        Link:
          $ref: '#/components/headers/Link'
        Content-Range:
          $ref: '#/components/headers/ContentRange'
...
    ApiDocs:
      description: The OpenAPI specification of this service.
      content:
        application/json:
          schema:
            $ref: 'https://spec.openapis.org/oas/3.0/schema/2021-09-28'
        text/yaml:
          schema:
            $ref: 'https://spec.openapis.org/oas/3.0/schema/2021-09-28'
        text/html: { }
      headers:
        Link:
          $ref: '#/components/headers/Link'
...
    Badges:
      description: The badges.
      content:
        application/json:
          schema:
            $ref: 'schema/badges.json'
        text/yaml:
          schema:
            $ref: 'schema/badges.json'
    Announcements:
      description: The announcements.
      content:
        application/json:
          schema:
            $ref: 'schema/announcements.json'
        text/yaml:
          schema:
            $ref: 'schema/announcements.json'
    PrivacyStatements:
      description: The privacy statements.
      content:
        application/json:
          schema:
            $ref: 'schema/privacy-statements.json'
        text/yaml:
          schema:
            $ref: 'schema/privacy-statements.json'
...
    Users:
      description: The users.
      content:
        application/json:
          schema:
            $ref: 'schema/users.json'
        text/yaml:
          schema:
            $ref: 'schema/users.json'
      headers:
        Link:
          $ref: '#/components/headers/Link'
        Content-Range:
          $ref: '#/components/headers/ContentRange'
...
    Fuelings:
      description: The fuelings.
      content:
        application/json:
          schema:
            $ref: 'schema/fuelings.json'
        text/yaml:
          schema:
            $ref: 'schema/fuelings.json'
      headers:
        Link:
          $ref: '#/components/headers/Link'
        Content-Range:
          $ref: '#/components/headers/ContentRange'
    Groups:
      description: The groups.
      content:
        application/json:
          schema:
            $ref: 'schema/groups.json'
        text/yaml:
          schema:
            $ref: 'schema/groups.json'
      headers:
        Link:
          $ref: '#/components/headers/Link'
        Content-Range:
          $ref: '#/components/headers/ContentRange'
...
  schemas:
    TemporalFilter:
      oneOf:
        - type: string
          description: Two comma-seperated date-times that form a temporal interval.
        - format: date-time
          type: string
          description: The temporal instant.
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic